requirements.txt
asgiref==3.6.0
Django==3.1.3
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
pymongo==3.12.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.2.4
tzdata==2023.3
whitenoise==6.4.0
